Streamline handling of iterate events with & without sort / search
Removed SearchForFearuresOrderByEvent and incorporated it into IterateFeaturesEvent to merge execution paths of both completely separate feature-iteration implementations.
