+++ xyz-connectors/src/main/java/com/here/xyz/connectors/ErrorResponseException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2019 HERE Europe B.V.
+ * Copyright (C) 2017-2022 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,26 +23,31 @@ import com.here.xyz.responses.XyzError;
 import com.here.xyz.responses.ErrorResponse;
 
 /**
- * An exception, which contains an ErrorResponse object.
+ * An exception, which will cause the connector to respond with an ErrorResponse object.
  */
 public class ErrorResponseException extends Exception {
 
   private ErrorResponse errorResponse;
 
+  public ErrorResponseException(XyzError xyzError, String errorMessage) {
+    this(null, xyzError, errorMessage);
+  }
+
   public ErrorResponseException(String streamId, XyzError xyzError, String errorMessage) {
     super(errorMessage);
-    this.errorResponse = new ErrorResponse()
-        .withStreamId(streamId)
-        .withError(xyzError)
-        .withErrorMessage(errorMessage);
+    createErrorResponse(streamId, xyzError, errorMessage);
   }
 
   public ErrorResponseException(String streamId, XyzError xyzError, Exception e) {
     super(e);
+    createErrorResponse(streamId, xyzError, e.getMessage());
+  }
+
+  private void createErrorResponse(String streamId, XyzError xyzError, String errorMessage) {
     this.errorResponse = new ErrorResponse()
         .withStreamId(streamId)
         .withError(xyzError)
-        .withErrorMessage(e.getMessage());
+        .withErrorMessage(errorMessage);
   }
 
   public ErrorResponse getErrorResponse() {
