+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/query/GetFeaturesByBBox.java
@@ -2,6 +2,7 @@ package com.here.xyz.psql.query;
 
 import static com.here.xyz.events.ContextAwareEvent.SpaceContext.DEFAULT;
 
+import com.here.xyz.connectors.ErrorResponseException;
 import com.here.xyz.events.GetFeaturesByBBoxEvent;
 import com.here.xyz.models.geojson.coordinates.BBox;
 import com.here.xyz.psql.DatabaseHandler;
@@ -10,17 +11,18 @@ import java.sql.SQLException;
 
 public class GetFeaturesByBBox<E extends GetFeaturesByBBoxEvent> extends ExtendedSpace<E> {
 
-  public GetFeaturesByBBox(E event, DatabaseHandler dbHandler) throws SQLException {
+  public GetFeaturesByBBox(E event, DatabaseHandler dbHandler) throws SQLException, ErrorResponseException {
     super(event, dbHandler);
   }
 
   @Override
-  protected SQLQuery buildQuery(E event) throws SQLException {
+  protected SQLQuery buildQuery(E event) {
     //NOTE: So far this query runner only handles queries regarding extended spaces
     if (isExtendedSpace(event) && event.getContext() == DEFAULT) {
       String filterWhereClause = "ST_Intersects(geo, ST_MakeEnvelope(#{minLon}, #{minLat}, #{maxLon}, #{maxLat}, 4326))";
       SQLQuery query = buildExtensionQuery(event, filterWhereClause);
 
+      //query.setQueryFragment("filterWhereClause", filterWhereClause);
       final BBox bbox = event.getBbox();
       query.setNamedParameter("minLon", bbox.minLon());
       query.setNamedParameter("minLat", bbox.minLat());
