+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/query/GetFeaturesById.java
@@ -2,6 +2,7 @@ package com.here.xyz.psql.query;
 
 import static com.here.xyz.events.ContextAwareEvent.SpaceContext.DEFAULT;
 
+import com.here.xyz.connectors.ErrorResponseException;
 import com.here.xyz.events.GetFeaturesByIdEvent;
 import com.here.xyz.psql.DatabaseHandler;
 import com.here.xyz.psql.SQLQuery;
@@ -9,12 +10,12 @@ import java.sql.SQLException;
 
 public class GetFeaturesById extends ExtendedSpace<GetFeaturesByIdEvent> {
 
-  public GetFeaturesById(GetFeaturesByIdEvent event, DatabaseHandler dbHandler) throws SQLException {
+  public GetFeaturesById(GetFeaturesByIdEvent event, DatabaseHandler dbHandler) throws SQLException, ErrorResponseException {
     super(event, dbHandler);
   }
 
   @Override
-  protected SQLQuery buildQuery(GetFeaturesByIdEvent event) throws SQLException {
+  protected SQLQuery buildQuery(GetFeaturesByIdEvent event) {
     String[] idArray = event.getIds().toArray(new String[0]);
     String filterWhereClause = "jsondata->>'id' = ANY(#{ids})";
 
@@ -22,6 +23,7 @@ public class GetFeaturesById extends ExtendedSpace<GetFeaturesByIdEvent> {
         ? buildExtensionQuery(event, filterWhereClause)
         : buildQuery(event, filterWhereClause);
 
+    //query.setQueryFragment("filterWhereClause", filterWhereClause);
     query.setNamedParameter("ids", idArray);
     return query;
   }
